DelegateProgrammatically
Attempt to create a delegated scope programmatically and check that it is delegated. Compile copiedDelegation.c as for previous delegator.c and delegate.c--Will need functions from inside systemd to work, but basics of programmatic checking is there.

Also, moved delegator.c, which was uploaded in the wrong folder. No changes to it.