SupportSystemD
Three files, implementing limited support for use of SystemD and python bindings to those functions.
cgroup_create_scope_and_slice and cgroup_create_scope_user_slice create a scope in a custom slice or the user slice, respectively, and allow the caller to control delegation of scope (a yes/no choice) as well as unit mode (via defined values). Both will leave a permanent sleep process in the created scope, and return either an error code (a negative number, or 0) or the PID (positive) of the permasleep process.
cgroup_is_delegated determines if the given scope/slice is delegated, taking the full path to the desired SystemD unit.